Fetch and parse Reddit RSS feed
This commit adds a function to fetch an RSS (technically Atom) feed from
Reddit and parse it into an X-expression. The feed content will be dealt
with in another commit.

Also, I finally gave in and added the 'threading' library for the ~>
macro. I still don't want to deal with packages though, so I've just
copied the code into this repo.