Add a couple more Reddit RSS feeds
These are the feeds I was originally planning to use the Reddit JSON
source for. However, Reddit blocks requests from GitHub Actions, so that
won't work without a bypass for that block.

Additionally, this commit switches to using the "submitted" RSS feeds
from Reddit, which only include submissions (posts) and not comments.
While the #:posts-only? filter already serves this purpose, it won't
help if the original RSS feed is filled with comments.